@@ -5,8 +5,10 @@ namespace GhibliQL\Type;
 use GraphQL\Type\Definition\ObjectType;
 use GraphQL\Type\Definition\ResolveInfo;
 use GraphQL\Type\Definition\Type;
+use GhibliQL\AppContext;
 use GhibliQL\Types;
 use GhibliQL\Data\DataSource;
+use GhibliQL\Data\Vehicle;
 
 class VehicleType extends ObjectType
 {
@@ -65,7 +67,7 @@ class VehicleType extends ObjectType
         parent::__construct($config);
     }
 
-    public function films($value, $args, $context, ResolveInfo $info)
+    public function films(Vehicle $value, array $args, AppContext $context, ResolveInfo $info): mixed
     {
         $film = null;
 
@@ -78,7 +80,7 @@ class VehicleType extends ObjectType
         return $film;
     }
 
-    public function pilot($value, $args, $context, ResolveInfo $info)
+    public function pilot(Vehicle $value, array $args, AppContext $context, ResolveInfo $info): mixed
     {
         $people = null;
 
