@@ -11,57 +11,39 @@ use GhibliQL\Type\QueryType;
 
 class Types
 {
-    private static $query;
-    private static $film;
-    private static $people;
-    private static $specie;
-    private static $location;
-    private static $vehicle;
+    private static ?QueryType $query = null;
+    private static ?FilmType $film = null;
+    private static ?PeopleType $people = null;
+    private static ?SpecieType $specie = null;
+    private static ?LocationType $location = null;
+    private static ?VehicleType $vehicle = null;
 
-    /**
-     * @return FilmType
-     */
-    public static function film()
+    public static function film(): FilmType
     {
         return self::$film ?: (self::$film = new FilmType());
     }
 
-    /**
-     * @return PeopleType
-     */
-    public static function people()
+    public static function people(): PeopleType
     {
         return self::$people ?: (self::$people = new PeopleType());
     }
 
-    /**
-     * @return SpecieType
-     */
-    public static function specie()
+    public static function specie(): SpecieType
     {
         return self::$specie ?: (self::$specie = new SpecieType());
     }
 
-    /**
-     * @return LocationType
-     */
-    public static function location()
+    public static function location(): LocationType
     {
         return self::$location ?: (self::$location = new LocationType());
     }
 
-    /**
-     * @return VehicleType
-     */
-    public static function vehicle()
+    public static function vehicle(): VehicleType
     {
         return self::$vehicle ?: (self::$vehicle = new VehicleType());
     }
 
-    /**
-     * @return QueryType
-     */
-    public static function query()
+    public static function query(): QueryType
     {
         return self::$query ?: (self::$query = new QueryType());
     }
